Capability-driven routing: per-capability controllers + conditional wiring
Replaces the monolithic Controller (all services required, even when the
adapter doesn't implement them) with seven per-capability controllers,
each conditionally wired based on whether its service bean exists.

Principle: capability presence ↔ route presence, no exceptions.

  * An adapter that doesn't wire a TokenService doesn't expose /api/assets/*.
  * An adapter that doesn't wire a PlanApprovalService doesn't expose /api/plan/*.
  * An adapter with zero services starts cleanly and serves nothing — accurately
    reflecting that it has no capabilities to offer (no misleading stubs that
    return 500 / 501 at runtime instead of being absent at wiring time).

Skeleton:
  * Controllers split into TokenController / EscrowController / PaymentController /
    PlanApprovalController / ReceiptController / OperationStatusController /
    HealthController, living under `io.ownera.finp2p.skeleton.web` —
    deliberately OUTSIDE any adapter's typical @SpringBootApplication
    component-scan root, so they're never auto-detected.
  * SkeletonWebAutoConfiguration declares each controller as a @Bean with
    @ConditionalOnBean on its service. Auto-config runs AFTER user config, so
    the conditional fires reliably against the fully-populated context.
    Registered via META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports.
  * DefaultHealthService — no-op HealthService impl adapters can wire with a
    one-line @Bean when they want probes exposed but don't need custom logic.
  * Monolithic Controller.java deleted.

Tests (11 new, 151 across the reactor):
  * EmptyAdapterTest (7 tests) — Spring Boot context starts with zero service
    beans wired, and every framework route returns 404. Pinned per capability:
    plan, token, escrow, payment, receipt + operation status, and health.
  * ObserverAdapterTest (4 tests) — only PlanApprovalService is wired
    (limited-access deployment shape from the request that motivated the
    refactor); /api/plan/approve is active (200), everything else stays 404.
  * Existing 140 tests pass unchanged — same URLs, same shapes, same Spring
    wiring for adapters that provide the full set of services.
