--- skeleton/src/main/java/io/ownera/ledger/adapter/Controller.java
@@ -1,449 +1,0 @@
-package io.ownera.ledger.adapter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.ownera.ledger.adapter.api.model.*;
-import io.ownera.ledger.adapter.service.*;
-import io.ownera.ledger.adapter.service.model.*;
-import io.ownera.ledger.adapter.service.workflow.CorrelationIdGenerator;
-import io.ownera.ledger.adapter.service.workflow.OperationStore;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
-
-import static io.ownera.ledger.adapter.Mappers.*;
-
-@RestController
-public class Controller {
-
-    private final EscrowService escrowService;
-    private final TokenService tokenService;
-    private final PaymentService paymentService;
-    private final PlanApprovalService planApprovalService;
-    private final CommonService commonService;
-    private final HealthService healthService;
-    private final SignatureVerifier signatureVerifier;
-    private final Optional<TransactionHook> transactionHook;
-    private final OperationStore operationStore;
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    public Controller(EscrowService escrowService, TokenService tokenService, PaymentService paymentService,
-                      PlanApprovalService planApprovalService, CommonService commonService,
-                      HealthService healthService, SignatureVerifier signatureVerifier,
-                      Optional<TransactionHook> transactionHook,
-                      OperationStore operationStore) {
-        this.escrowService = escrowService;
-        this.tokenService = tokenService;
-        this.paymentService = paymentService;
-        this.planApprovalService = planApprovalService;
-        this.commonService = commonService;
-        this.healthService = healthService;
-        this.signatureVerifier = signatureVerifier;
-        this.transactionHook = transactionHook;
-        this.operationStore = operationStore;
-    }
-
-    private final static Logger logger = LoggerFactory.getLogger(Controller.class);
-
-    private static String ensureIdempotencyKey(String idempotencyKey) {
-        return idempotencyKey != null ? idempotencyKey : CorrelationIdGenerator.generate();
-    }
-
-    // --- Health endpoints ---
-
-    @GetMapping(value = "/health")
-    public ResponseEntity<String> health() {
-        return ResponseEntity.ok("OK");
-    }
-
-    @GetMapping(value = "/health/liveness")
-    public ResponseEntity<String> liveness(
-            @RequestHeader(name = "skip-vendor", required = false) String skipVendor
-    ) {
-        if (!"true".equals(skipVendor)) {
-            healthService.liveness();
-        }
-        return ResponseEntity.ok("OK");
-    }
-
-    @GetMapping(value = "/health/readiness")
-    public ResponseEntity<String> readiness(
-            @RequestHeader(name = "skip-vendor", required = false) String skipVendor
-    ) {
-        if (!"true".equals(skipVendor)) {
-            healthService.readiness();
-        }
-        return ResponseEntity.ok("OK");
-    }
-
-    // --- Execution plan endpoints ---
-
-    @PostMapping(value = "/api/plan/approve", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIApproveExecutionPlanResponse> approvePlan(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIApproveExecutionPlanRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        String planId = request.getExecutionPlan().getId();
-        logger.info("Approve plan: {}", planId);
-        PlanApprovalStatus status = planApprovalService.approvePlan(ik, planId);
-        return ResponseEntity.status(HttpStatus.OK).body(toAPIResponse(status));
-    }
-
-    @PostMapping(value = "/api/plan/proposal", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIApproveExecutionPlanResponse> planProposal(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIExecutionPlanProposalRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        String planId = request.getExecutionPlan().getId();
-        Object proposal = request.getExecutionPlan().getProposal().getActualInstance();
-        logger.info("Plan proposal: planId={}, type={}", planId, proposal.getClass().getSimpleName());
-
-        PlanApprovalStatus status;
-        if (proposal instanceof APIExecutionPlanCancellationProposal) {
-            status = planApprovalService.proposeCancelPlan(ik, planId);
-        } else if (proposal instanceof APIExecutionPlanResetProposal) {
-            APIExecutionPlanResetProposal reset = (APIExecutionPlanResetProposal) proposal;
-            int seq = reset.getProposedSequence() != null ? reset.getProposedSequence() : 0;
-            status = planApprovalService.proposeResetPlan(ik, planId, seq);
-        } else if (proposal instanceof APIExecutionPlanInstructionProposal) {
-            APIExecutionPlanInstructionProposal instr = (APIExecutionPlanInstructionProposal) proposal;
-            int seq = instr.getInstructionSequence() != null ? instr.getInstructionSequence() : 0;
-            status = planApprovalService.proposeInstructionApproval(ik, planId, seq);
-        } else {
-            logger.warn("Unknown proposal type: {}", proposal.getClass().getName());
-            status = new ApprovedPlan();
-        }
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPIResponse(status));
-    }
-
-    @PostMapping(value = "/api/plan/proposal/status", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Void> proposalStatus(
-            @RequestBody APIExecutionPlanProposalStatusRequest request
-    ) {
-        String planId = request.getRequest().getExecutionPlan().getId();
-        PlanProposal proposal = proposalFromAPI(request.getRequest().getExecutionPlan().getProposal());
-        ProposalStatus status = "rejected".equalsIgnoreCase(
-                request.getStatus() != null ? request.getStatus().getValue() : null)
-                ? ProposalStatus.REJECTED : ProposalStatus.APPROVED;
-        logger.info("Proposal status: planId={}, status={}, type={}", planId, status, proposal.getClass().getSimpleName());
-        planApprovalService.proposalStatus(planId, proposal, status);
-        return ResponseEntity.noContent().build();
-    }
-
-    // --- Asset management endpoints ---
-
-    @PostMapping(value = "/api/assets/create", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APICreateAssetResponse> createAsset(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APICreateAssetRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Create asset: {}", request);
-        AssetCreationStatus status = tokenService.createAsset(
-                ik,
-                fromAPI(request.getAsset()),
-                fromAPI(request.getLedgerAssetBinding()),
-                request.getMetadata(),
-                request.getName(),
-                request.getIssuerId(),
-                fromAPI(request.getDenomination())
-        );
-        return ResponseEntity.status(HttpStatus.OK).body(toAPIResponse(status));
-    }
-
-    // --- Token operation endpoints ---
-
-    @PostMapping(value = "/api/assets/issue", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> issue(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIIssueAssetsRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Issue assets: {}", request);
-
-        FinIdAccount destination = finIdAccountFromAPI(request.getDestination());
-        Asset asset = assetFromAPI(request.getDestination());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.ISSUE, null,
-                destination.destination(), asset, request.getQuantity(), null, exCtx));
-        ReceiptOperation rcptOp = tokenService.issue(ik, asset, destination, request.getQuantity(), exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.ISSUE, null,
-                destination.destination(), asset, request.getQuantity(), null, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    @PostMapping(value = "/api/assets/transfer", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> transfer(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APITransferAssetRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Transfer assets: {}", request);
-
-        Source source = sourceFromAPI(request.getSource());
-        Destination destination = destinationFromAPI(request.getDestination());
-        Asset asset = assetFromAPI(request.getSource());
-        Signature sig = fromAPI(request.getSignature());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        // Uncomment to enable signature verification:
-        // if (!signatureVerifier.verify(sig, request.getSource().getFinId())) {
-        //     throw new BusinessException(4, "Signature verification failed");
-        // }
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.TRANSFER, source, destination, asset,
-                request.getQuantity(), sig, exCtx));
-        ReceiptOperation rcptOp = tokenService.transfer(ik, request.getNonce(), source, destination, asset,
-                request.getQuantity(), sig, exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.TRANSFER, source, destination, asset,
-                request.getQuantity(), sig, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    @PostMapping(value = "/api/assets/redeem", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> redeem(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIRedeemAssetsRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Redeem assets: {}", request);
-
-        FinIdAccount source = finIdAccountFromAPI(request.getSource());
-        Asset asset = assetFromAPI(request.getSource());
-        Signature sig = fromAPI(request.getSignature());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        // Uncomment to enable signature verification:
-        // if (!signatureVerifier.verify(sig, request.getSource().getFinId())) {
-        //     throw new BusinessException(4, "Signature verification failed");
-        // }
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.REDEEM, source.source(), null, asset,
-                request.getQuantity(), sig, exCtx));
-        ReceiptOperation rcptOp = tokenService.redeem(ik, request.getNonce(), source, asset,
-                request.getQuantity(), request.getOperationId(), sig, exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.REDEEM, source.source(), null, asset,
-                request.getQuantity(), sig, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    // --- Balance endpoints ---
-
-    @PostMapping(value = "/api/assets/getBalance", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIGetAssetBalanceResponse> getBalance(@RequestBody APIGetAssetBalanceRequest request) {
-        // 0.28: asset is embedded in the APIAccount (owner)
-        Asset asset = assetFromAPI(request.getOwner());
-        String balance = tokenService.getBalance(asset, request.getOwner().getFinId());
-        return ResponseEntity.status(HttpStatus.OK).body(new APIGetAssetBalanceResponse()
-                .asset(request.getOwner().getAsset())
-                .balance(balance));
-    }
-
-    @PostMapping(value = "/api/asset/balance", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIAssetBalanceInfoResponse> assetBalance(@RequestBody APIAssetBalanceInfoRequest request) {
-        Asset asset = fromAPI(request.getAsset());
-        String finId = request.getAccount().getFinId();
-        Balance balance = tokenService.balance(asset, finId);
-        return ResponseEntity.status(HttpStatus.OK).body(balanceToAPI(request.getAsset(), request.getAccount(), balance));
-    }
-
-    // --- Receipt & operation status endpoints ---
-
-    @GetMapping(value = "/api/assets/receipts/{id}")
-    public ResponseEntity<APIGetReceiptResponse> getReceipt(
-            @PathVariable("id") String transactionId
-    ) {
-        ReceiptOperation receipt = commonService.getReceipt(transactionId);
-        return ResponseEntity.status(HttpStatus.OK).body(Mappers.toAPIGetReceiptResponse(receipt));
-    }
-
-    @GetMapping(value = "/api/operations/status/{id}")
-    public final ResponseEntity<APIOperationStatus> getOperationStatus(@PathVariable("id") String correlationId) {
-        // Operations table is the source of truth for cid lookups (matches the Node
-        // workflow proxy in skeleton/src/workflows/service.ts: createServiceProxy
-        // returns operation.outputs directly). Reading outputs verbatim is the only
-        // correct path — CommonService.operationStatus on the underlying adapter
-        // looks up by transaction id, not by cid, so it is never appropriate here.
-        String stored = operationStore.findOutputsByCid(correlationId);
-        if (stored == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        try {
-            APIOperationStatus api = objectMapper.readValue(stored, APIOperationStatus.class);
-            return ResponseEntity.status(HttpStatus.OK).body(api);
-        } catch (Exception e) {
-            logger.error("Failed to parse persisted outputs for cid={}: {}", correlationId, e.getMessage());
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-    }
-
-    // --- Escrow endpoints ---
-
-    @PostMapping(value = "/api/assets/hold", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> hold(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIHoldOperationRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Hold assets: {}", request);
-
-        Source source = sourceFromAPI(request.getSource());
-        Destination destination = destinationFromAPI(request.getDestination());
-        Asset asset = assetFromAPI(request.getSource());
-        Signature sig = fromAPI(request.getSignature());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        // Uncomment to enable signature verification:
-        // if (!signatureVerifier.verify(sig, request.getSource().getFinId())) {
-        //     throw new BusinessException(4, "Signature verification failed");
-        // }
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.HOLD, source, destination, asset,
-                request.getQuantity(), sig, exCtx));
-        ReceiptOperation rcptOp = escrowService.hold(ik, request.getNonce(), source, destination, asset,
-                request.getQuantity(), sig, request.getOperationId(), exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.HOLD, source, destination, asset,
-                request.getQuantity(), sig, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    @PostMapping(value = "/api/assets/release", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> release(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIReleaseOperationRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Release assets: {}", request);
-
-        Source source = sourceFromAPI(request.getSource());
-        Destination destination = destinationFromAPI(request.getDestination());
-        Asset asset = assetFromAPI(request.getSource());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.RELEASE, source, destination, asset,
-                request.getQuantity(), null, exCtx));
-        ReceiptOperation rcptOp = escrowService.release(ik, source, destination, asset,
-                request.getQuantity(), request.getOperationId(), exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.RELEASE, source, destination, asset,
-                request.getQuantity(), null, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    @PostMapping(value = "/api/assets/rollback", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIReceiptOperation> rollback(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIRollbackOperationRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Rollback assets: {}", request);
-
-        Source source = sourceFromAPI(request.getSource());
-        Asset asset = assetFromAPI(request.getSource());
-        ExecutionContext exCtx = fromAPI(request.getExecutionContext());
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.ROLLBACK, source, null, asset,
-                request.getQuantity(), null, exCtx));
-        ReceiptOperation rcptOp = escrowService.rollback(ik, source, asset,
-                request.getQuantity(), request.getOperationId(), exCtx);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.ROLLBACK, source, null, asset,
-                request.getQuantity(), null, exCtx, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPI(rcptOp));
-    }
-
-    // --- Payment endpoints ---
-
-    @PostMapping(value = "/api/payments/depositInstruction", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIDepositInstructionResponse> depositInstruction(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIDepositInstructionRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Deposit instruction: {}", request);
-
-        // Uncomment to enable signature verification:
-        // Signature sig = fromAPI(request.getSignature());
-        // if (sig != null && !signatureVerifier.verify(sig, request.getOwner().getFinId())) {
-        //     throw new BusinessException(4, "Signature verification failed");
-        // }
-        DepositOperation rcptOp = paymentService.getDepositInstruction(
-                ik,
-                sourceFromAPI(request.getOwner()),
-                destinationFromAPI(request.getDestination()),
-                fromAPI(request.getAsset()),
-                request.getAmount(),
-                request.getDetails(),
-                request.getNonce(),
-                fromAPI(request.getSignature())
-        );
-        return ResponseEntity.status(HttpStatus.OK).body(toAPIResponse(rcptOp));
-    }
-
-    @PostMapping(value = "/api/payments/payout", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<APIPayoutResponse> payout(
-            @RequestHeader(name = "Idempotency-Key", required = false) String idempotencyKey,
-            @RequestBody APIPayoutRequest request
-    ) {
-        String ik = ensureIdempotencyKey(idempotencyKey);
-        logger.info("Payout: {}", request);
-
-        Source source = sourceFromAPI(request.getSource());
-        Destination destination = destinationFromAPI(request.getDestination());
-        Asset asset = fromAPI(request.getAsset()); // payout: asset is top-level APIPayoutAsset
-        Signature sig = fromAPI(request.getSignature());
-
-        // Uncomment to enable signature verification:
-        // if (sig != null && !signatureVerifier.verify(sig, request.getSource().getFinId())) {
-        //     throw new BusinessException(4, "Signature verification failed");
-        // }
-
-        transactionHook.ifPresent(h -> h.preTransaction(
-                ik, OperationType.TRANSFER, source, destination, asset,
-                request.getQuantity(), sig, null));
-        String description = null;
-        if (request.getPayoutInstruction() != null) {
-            description = request.getPayoutInstruction().getDescription();
-        }
-        String desc = description;
-        ReceiptOperation rcptOp = paymentService.payout(ik, source, destination, asset,
-                request.getQuantity(), desc, request.getNonce(), sig);
-
-        transactionHook.ifPresent(h -> h.postTransaction(
-                ik, OperationType.TRANSFER, source, destination, asset,
-                request.getQuantity(), sig, null, rcptOp));
-
-        return ResponseEntity.status(HttpStatus.OK).body(toAPIPayoutResponse(rcptOp));
-    }
-
-}
